+++ test/webpack/stripes-config-plugin.spec.js
@@ -18,6 +18,23 @@ const compilerStub = {
     },
     plugins: [],
   },
+  hooks: {
+    afterPlugins: {
+      tap: () => {},
+    },
+    emit: {
+      tapAsync: () => {},
+    },
+    afterEnvironment: {
+      tap: () => {}
+    },
+    afterResolvers: {
+      tap: () => {}
+    },
+    watchRun: {
+      tapAsync: () => {}
+    },
+  },
   context: '/context/path',
   warnings: [],
 };
@@ -59,19 +76,22 @@ describe('The stripes-config-plugin', function () {
       this.sut = new StripesConfigPlugin(mockConfig);
     });
 
+    afterEach(function () {
+      delete compilerStub.hooks.stripesConfigPluginBeforeWrite;
+    });
+
     it('applies a virtual module', function () {
-      this.sandbox.spy(compilerStub, 'apply');
+      this.sandbox.spy(VirtualModulesPlugin.prototype, 'apply');
       this.sut.apply(compilerStub);
 
-      expect(compilerStub.apply).to.have.been.calledOnce;
-      const applyCall = compilerStub.apply.getCall(0);
-      expect(applyCall.args[0]).to.be.an.instanceOf(VirtualModulesPlugin);
+      expect(VirtualModulesPlugin.prototype.apply).to.have.been.calledOnce;
+      expect(VirtualModulesPlugin.prototype.apply).to.be.calledWith(compilerStub);
     });
 
     it('registers the "after-plugins" hook', function () {
-      this.sandbox.spy(compilerStub, 'plugin');
+      this.sandbox.spy(compilerStub.hooks.afterPlugins, 'tap');
       this.sut.apply(compilerStub);
-      expect(compilerStub.plugin).to.have.been.calledWith('after-plugins');
+      expect(compilerStub.hooks.afterPlugins.tap).to.have.been.calledWith('StripesConfigPlugin');
     });
   });
 
@@ -93,6 +113,10 @@ describe('The stripes-config-plugin', function () {
       compilerStub.options.plugins.push(brandingPlugin);
 
       this.sut.apply(compilerStub);
+    });
+
+    afterEach(function () {
+      delete compilerStub.hooks.stripesConfigPluginBeforeWrite;
     });
 
     it('calls virtualModule.writeModule()', function () {