+++ webpack.config.base.js
@@ -5,7 +5,6 @@ const HtmlWebpackPlugin = require('html-webpack-plugin');
 const { generateStripesAlias } = require('./webpack/module-paths');
 const babelLoaderRule = require('./webpack/babel-loader-rule');
 const typescriptLoaderRule = require('./webpack/typescript-loader-rule');
-const StripesDuplicatesPlugin = require('./webpack/stripes-duplicate-plugin');
 
 // React doesn't like being included multiple times as can happen when using
 // yarn link. Here we find a more specific path to it by first looking in
@@ -29,16 +28,11 @@ module.exports = {
       template: `${__dirname}/index.html`,
     }),
     new webpack.EnvironmentPlugin(['NODE_ENV']),
-    new StripesDuplicatesPlugin(),
   ],
   module: {
     rules: [
       babelLoaderRule,
       typescriptLoaderRule,
-      {
-        test: /\.json$/,
-        loader: 'json-loader',
-      },
       {
         test: /\.(jpg|jpeg|gif|png|ico|svg)$/,
         loader: 'file-loader?name=img/[path][name].[hash].[ext]',