+++ webpack.config.cli.prod.js
@@ -1,7 +1,7 @@
 // Top level Webpack configuration for building static files for
 // production deployment from the command line
 
-const ExtractTextPlugin = require('extract-text-webpack-plugin');
+const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 const OptimizeCssAssetsPlugin = require('optimize-css-assets-webpack-plugin');
 const postCssImport = require('postcss-import');
 const autoprefixer = require('autoprefixer');
@@ -15,44 +15,46 @@ const postCssColorFunction = require('postcss-color-function');
 const base = require('./webpack.config.base');
 const cli = require('./webpack.config.cli');
 
-const prodConfig = Object.assign({}, base, cli);
+const prodConfig = Object.assign({}, base, cli, {
+  mode: 'production',
+});
 
 prodConfig.plugins = prodConfig.plugins.concat([
-  new ExtractTextPlugin({ filename: 'style.[contenthash].css', allChunks: true }),
+  new MiniCssExtractPlugin({ filename: 'style.[contenthash].css', allChunks: true }),
   new OptimizeCssAssetsPlugin(),
 ]);
 
 prodConfig.module.rules.push({
   test: /\.css$/,
-  use: ExtractTextPlugin.extract({
-    fallback: 'style-loader',
-    use: [
-      {
-        loader: 'css-loader',
-        options: {
-          localIdentName: '[local]---[hash:base64:5]',
-          modules: true,
-          importLoaders: 1,
-        },
+  use: [
+    {
+      loader: MiniCssExtractPlugin.loader,
+    },
+    {
+      loader: 'css-loader',
+      options: {
+        localIdentName: '[local]---[hash:base64:5]',
+        modules: true,
+        importLoaders: 1,
       },
-      {
-        loader: 'postcss-loader',
-        options: {
-          ident: 'postcss',
-          plugins: () => [
-            postCssImport(),
-            autoprefixer(),
-            postCssCustomProperties(),
-            postCssCalc(),
-            postCssNesting(),
-            postCssCustomMedia(),
-            postCssMediaMinMax(),
-            postCssColorFunction(),
-          ],
-        },
+    },
+    {
+      loader: 'postcss-loader',
+      options: {
+        ident: 'postcss',
+        plugins: () => [
+          postCssImport(),
+          autoprefixer(),
+          postCssCustomProperties(),
+          postCssCalc(),
+          postCssNesting(),
+          postCssCustomMedia(),
+          postCssMediaMinMax(),
+          postCssColorFunction(),
+        ],
       },
-    ],
-  }),
+    },
+  ],
 });
 
 // Remove all data-test or data-test-* attributes