+++ webpack/build.js
@@ -1,9 +1,6 @@
 const webpack = require('webpack');
 const path = require('path');
-const StripesConfigPlugin = require('./stripes-config-plugin');
-const StripesBrandingPlugin = require('./stripes-branding-plugin');
-const StripesTranslationsPlugin = require('./stripes-translations-plugin');
-const UglifyJSPlugin = require('uglifyjs-webpack-plugin');
+const StripesWebpackPlugin = require('./stripes-webpack-plugin');
 const applyWebpackOverrides = require('./apply-webpack-overrides');
 const logger = require('./logger')();
 
@@ -14,9 +11,7 @@ module.exports = function build(stripesConfig, options) {
     logger.log('starting build...');
     let config = require('../webpack.config.cli.prod'); // eslint-disable-line
 
-    config.plugins.push(new StripesConfigPlugin(stripesConfig));
-    config.plugins.push(new StripesBrandingPlugin(stripesConfig.branding));
-    config.plugins.push(new StripesTranslationsPlugin(stripesConfig));
+    config.plugins.push(new StripesWebpackPlugin({ stripesConfig }));
 
     config.resolve.modules = ['node_modules', platformModulePath];
     config.resolveLoader = { modules: ['node_modules', platformModulePath] };
@@ -30,10 +25,12 @@ module.exports = function build(stripesConfig, options) {
     if (options.sourcemap) {
       config.devtool = 'source-map';
     }
-    if (options.minify !== false) {
-      config.plugins.push(new UglifyJSPlugin({
-        sourceMap: config.devtool && config.devtool === 'source-map',
-      }));
+
+    // By default, Webpack's production mode will configure UglifyJS
+    // Override this when we explicity set --no-minify on the command line
+    if (options.minify === false) {
+      config.optimization = config.optimization || {};
+      config.optimization.minimize = false;
     }
 
     // Give the caller a chance to apply their own webpack overrides